+++ lnwire/lnwire.go
@@ -62,13 +62,13 @@ func writeElement(w io.Writer, element interface{}) error {
 	switch e := element.(type) {
 	case uint8:
 		var b [1]byte
-		b[0] = byte(e)
+		b[0] = e
 		if _, err := w.Write(b[:]); err != nil {
 			return err
 		}
 	case uint16:
 		var b [2]byte
-		binary.BigEndian.PutUint16(b[:], uint16(e))
+		binary.BigEndian.PutUint16(b[:], e)
 		if _, err := w.Write(b[:]); err != nil {
 			return err
 		}
@@ -86,13 +86,13 @@ func writeElement(w io.Writer, element interface{}) error {
 		}
 	case uint32:
 		var b [4]byte
-		binary.BigEndian.PutUint32(b[:], uint32(e))
+		binary.BigEndian.PutUint32(b[:], e)
 		if _, err := w.Write(b[:]); err != nil {
 			return err
 		}
 	case uint64:
 		var b [8]byte
-		binary.BigEndian.PutUint64(b[:], uint64(e))
+		binary.BigEndian.PutUint64(b[:], e)
 		if _, err := w.Write(b[:]); err != nil {
 			return err
 		}
